# src/v3/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nova API — v3",
    "version": "3.0.0",
    "description": "API v3 - Condomínio Nova Residence"
  },
...
  "components": {
    "schemas": {
      "ResponseMeta": {
        "type": "object",
        "description": "Metadado transversal presente em toda resposta, dentro ou fora do envelope de sucesso/erro.",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Instante em que a resposta foi gerada.",
            "example": "2026-09-28T03:44:40.342Z"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "description": "Identificador da requisição, para correlação em logs e suporte.",
            "example": "req-1"
          },
          "version": {
            "type": "string",
            "enum": ["v3"],
            "description": "Versão da API que gerou a resposta."
          },
          "pagination": {
            "type": "object",
            "description": "Presente somente em endpoints que retornam uma coleção paginada.",
            "properties": {
              "limit": { "type": "integer", "minimum": 1, "example": 25 },
              "offset": { "type": "integer", "minimum": 0, "example": 0 },
              "total": { "type": "integer", "minimum": 0, "example": 137 }
            },
            "required": ["limit", "offset", "total"]
          }
        },
        "required": ["timestamp", "requestId", "version"]
      },
      "ProblemDetails": {
        "type": "object",
        "description": "Estrutura padrão de erro da API, com tipo, título, status HTTP e detalhe da ocorrência.",
        "properties": {
          "type": {
            "type": "string",
            "description": "URN fixo identificando o tipo do erro (não é uma URL navegável).",
            "example": "urn:nova-api:validation-error"
          },
          "title": {
            "type": "string",
            "description": "Resumo legível do tipo de erro.",
            "example": "Dados de entrada inválidos"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP correspondente, redundante com o status da resposta.",
            "example": 400
          },
          "detail": {
            "type": "string",
            "nullable": true,
            "description": "Explicação específica desta ocorrência do erro.",
            "example": "Um ou mais campos da requisição são inválidos."
          },
          "instance": {
            "type": "string",
            "nullable": true,
            "description": "Path da requisição que originou o erro.",
            "example": "/v3/api/health"
          },
          "validationErrors": {
            "type": "array",
            "description": "Presente somente quando type = validation-error.",
            "items": {
              "type": "object",
              "properties": {
                "path": { "type": "string", "example": "/name" },
                "message": { "type": "string", "example": "Too small: expected string to have >=3 characters" }
              },
              "required": ["path", "message"]
            }
          }
        },
        "required": ["type", "title", "status", "detail", "instance"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "$ref": "#/components/schemas/ProblemDetails" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "required": ["success", "error", "meta"]
      },
...
      "ResidenceUnit": {
        "type": "object",
        "properties": {
          "sequencia": { "type": "integer" },
          "quadra": { "type": "string" },
          "lote": { "type": "string" },
          "label": { "type": "string", "example": "A-124" },
          "morador": { "type": "boolean" },
          "proprietario": { "type": "boolean" },
          "locatario": { "type": "boolean" },
          "blocked": { "type": "boolean", "description": "true = mostrar a unidade, mas desativar toda ação nela." },
          "blockedReason": {
            "type": "string",
            "nullable": true,
            "enum": ["nao-morador", "unidade-com-inquilino", "cadastro-inconsistente"],
            "description": "nao-morador: MOR diferente de S. unidade-com-inquilino: proprietário numa unidade com locatário (lei do inquilinato). cadastro-inconsistente: PROP e LOC juntos, ou nenhum dos dois — corrigir o cadastro."
          }
        },
        "required": [
          "sequencia",
          "quadra",
          "lote",
          "label",
          "morador",
          "proprietario",
          "locatario",
          "blocked",
          "blockedReason"
        ]
      },
      "Residence": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["pendente", "ativo", "sem-unidade"],
            "description": "pendente: conta ainda não vinculada ao cadastro. sem-unidade: vinculada, sem ligação com unidade no momento."
          },
          "blocked": {
            "type": "boolean",
            "description": "true quando nenhuma unidade está liberada — app todo desativado. O login continua funcionando."
          },
          "person": {
            "type": "object",
            "nullable": true,
            "properties": { "sequencia": { "type": "integer" }, "nome": { "type": "string" } }
          },
          "units": { "type": "array", "items": { "$ref": "#/components/schemas/ResidenceUnit" } }
        },
        "required": ["status", "blocked", "person", "units"]
      },
      "AccountLink": {
        "type": "object",
        "properties": {
          "accountId": { "type": "string", "format": "uuid" },
          "personSequencia": { "type": "integer", "description": "PESSOAS.SEQUENCIA no Firebird." },
          "linkedBy": { "type": "string", "format": "uuid" },
          "linkedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["accountId", "personSequencia", "linkedBy", "linkedAt"]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado — token ausente, inválido ou expirado.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token obtido em POST /v3/api/auth/login. Expira em 15 minutos; renovar em POST /v3/api/auth/refresh."
      }
    }
  },
  "paths": {
...
    "/v3/api/residence/me": {
      "get": {
        "tags": ["Residência"],
        "summary": "Situação residencial da conta",
        "description": "Chamar logo após o login. Define o que o app libera: unidades vinculadas no cadastro do condomínio e se cada uma está bloqueada. Consulta o Firebird com cache de 2 minutos.",
        "operationId": "getMyResidence",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Situação atual.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "data": { "$ref": "#/components/schemas/Residence" },
                    "meta": { "$ref": "#/components/schemas/ResponseMeta" }
                  },
                  "required": ["success", "data", "meta"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v3/api/residence/links": {
      "get": {
        "tags": ["Residência"],
        "summary": "Listar vínculos (síndico/admin)",
        "operationId": "listResidenceLinks",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Vínculos existentes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/AccountLink" } },
                    "meta": { "$ref": "#/components/schemas/ResponseMeta" }
                  },
                  "required": ["success", "data", "meta"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Papel sem permissão para esta rota.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/v3/api/residence/links/{accountId}": {
      "put": {
        "tags": ["Residência"],
        "summary": "Vincular conta a pessoa (síndico/admin)",
        "description": "Liga a conta ao cadastro PESSOAS.SEQUENCIA do Firebird. Uma pessoa só pode estar vinculada a uma conta.",
        "operationId": "linkResidence",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "description": "Id da conta (auth).",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "personSequencia": { "type": "integer" } },
                "required": ["personSequencia"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vínculo criado ou trocado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "data": { "$ref": "#/components/schemas/AccountLink" },
                    "meta": { "$ref": "#/components/schemas/ResponseMeta" }
                  },
                  "required": ["success", "data", "meta"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Papel sem permissão para esta rota.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Pessoa não encontrada no cadastro.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "409": {
            "description": "Pessoa já vinculada a outra conta.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["Residência"],
        "summary": "Remover vínculo (síndico/admin)",
        "operationId": "unlinkResidence",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "description": "Id da conta (auth).",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Vínculo removido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "data": { "nullable": true },
                    "meta": { "$ref": "#/components/schemas/ResponseMeta" }
                  },
                  "required": ["success", "data", "meta"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Papel sem permissão para esta rota.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Conta sem vínculo.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}